+++ src/main/java/com/BackEnd/BidPro/Repo/CategoryRepo.java
@@ -3,5 +3,5 @@ package com.BackEnd.BidPro.Repo;
 import com.BackEnd.BidPro.Model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepo extends JpaRepository<Category, Integer> {
+public interface CategoryRepo extends JpaRepository<Category, Long> {
 }
